create TrinketCarousel component and initial set up
