+++ src/home/middle/Market.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { useSpring, animated } from "react-spring"
 import RegionCarousel from './RegionCarousel'
+import TrinketCarousel from './TrinketCarousel'
 
 const Market = () => {
 
@@ -14,6 +15,7 @@ const Market = () => {
     return (
         <animated.div style={fadeIn} className="market">
             <RegionCarousel />
+            <TrinketCarousel />
         </animated.div>
     )
 }
